+++ tests/test_electrodes.py
@@ -1,25 +1,198 @@
-# import pytest
+import pytest
 import numpy as np
 from numpy.testing import assert_allclose
 
+import emg3d
 from emg3d import electrodes
 
 
-def test_inst_to_from_dict():
+class TestWire:
 
-    # RxElectricPoint
-    r1a = electrodes.RxElectricPoint((1200, -56, 23.214, 368, 15))
-    r1b = electrodes.RxElectricPoint.from_dict(r1a.to_dict())
-    assert r1a == r1b
+    class Magnetic(electrodes.Dipole):
+        def __init__(self, coordinates, *args, **kwargs):
+            self._coordinates = coordinates
+            super().__init__(coordinates, *args, **kwargs)
 
-    # RxMagneticPoint
-    r2a = electrodes.RxMagneticPoint((-1200, 56, -23.214, 0, 90))
-    r2b = electrodes.RxMagneticPoint.from_dict(r2a.to_dict())
-    assert r2a == r2b
+    class Electric(electrodes.Dipole):
+        def __init__(self, coordinates, *args, **kwargs):
+            self._coordinates = coordinates
+            super().__init__(coordinates, *args, **kwargs)
 
-    # TxElectricDipole - 3 formats
+    def test_basics(self):
+
+        p1 = [[0, 0, 0], [1, 1, 1]]
+        p2 = [0.5, 0.5, 0.5]
+
+        e1 = electrodes.Wire(p1)
+        e2 = electrodes.Wire(p2)
+
+        assert e1 == e1
+        assert e1 != e2
+        assert_allclose(e1.points, e1.coordinates)
+        assert e1.xtype == 'electric'
+        assert_allclose(e2.center, p2)
+        assert e2.length == 0
+        assert e1.segment_n == 1
+        assert e2.segment_n == 0
+        assert e1.segment_lengths == np.sqrt(3)
+        assert np.empty(e2.segment_lengths)
+
+        e3 = e1.copy()
+        assert e1 == e3
+
+        de3 = e3.to_dict()
+        assert_allclose(de3['coordinates'], p1)
+
+        # Dummy test
+        e4 = self.Magnetic(p1)
+        assert e4.xtype == 'magnetic'
+        assert e4.coordinates.shape != e4.points.shape
+
+        e5 = self.Electric(p1)
+        assert e5.xtype == 'electric'
+        assert_allclose(e5.coordinates, e5.points)
+
+    def test_basic_repr(self):
+        p1 = np.array([[-10, -10, -10], [10, 10, 10]])
+        e1 = electrodes.Wire(p1)
+        r1 = e1.__repr__()
+        assert 'Wire' in r1
+        assert 'center={0.0; 0.0; 0.0} m' in r1
+        assert 'n=1; l=' in r1
+
+    def test_warnings(self):
+
+        # Too long 1D array.
+        with pytest.raises(ValueError, match="`coordinates` must be of shape"):
+            electrodes.Wire((0, 1, 2, 3, 4, 5, 6))
+
+        # 3D array.
+        with pytest.raises(ValueError, match="`coordinates` must be of shape"):
+            electrodes.Wire(np.ones((3, 3, 3)))
+
+
+def test_point():
+    p1 = (100, 0, -20, 15, 70.1234)
+    e1 = electrodes.Point(p1)
+    assert_allclose(e1.points, [p1[:3]])
+    assert_allclose(e1.center, p1[:3])
+    assert e1.azimuth == p1[3]
+    assert e1.elevation == p1[4]
+
+    r1 = e1.__repr__()
+    assert 'Point' in r1
+    assert 'x=100.0 m, y=0.0 m, z=-20.0 m' in r1
+    assert '15.0°' in r1
+    assert '70.1°' in r1
+
+    with pytest.raises(ValueError, match="Point coordinates are wrong "):
+        electrodes.Point((0, 1, 2, 3, 4, 5, 6))
+
+
+class TestDipole:
+
+    class Magnetic(electrodes.Dipole):
+        def __init__(self, *args, **kwargs):
+            super().__init__(*args, **kwargs)
+
+    def test_point(self):
+        p1 = (100, 2, 6, 0, 90)
+        e1 = electrodes.Dipole(p1)
+        assert e1.length == 1.0
+        assert_allclose(p1, e1.coordinates)
+        assert_allclose([p1[0], p1[1], p1[2]-0.5], e1.points[0, :])
+        assert e1.azimuth == 0
+        assert e1.elevation == 90
+
+        e2 = electrodes.Dipole(p1, np.pi)
+        assert e2.length == np.pi
+
+        # Dummy test
+        e3 = self.Magnetic(p1)
+        assert e3.xtype == 'magnetic'
+        assert e3.points.shape == (5, 3)
+
+    def test_flat(self):
+        p1 = (-10, 10, -10, 10, -10, 10)
+        e1 = electrodes.Dipole(p1)
+        assert_allclose(e1.length, 20*np.sqrt(3))
+        assert_allclose(p1, e1.coordinates)
+        assert_allclose(p1[::2], e1.points[0, :])
+        assert e1.elevation == 35.264389682754654
+        assert e1.azimuth == 45
+
+    def test_dipole(self):
+        p1 = np.array([[-10, -10, -10], [10, 10, 10]])
+        e1 = electrodes.Dipole(p1)
+        assert_allclose(e1.length, 20*np.sqrt(3))
+        assert_allclose(p1, e1.coordinates)
+        assert_allclose(p1, e1.points)
+        assert e1.azimuth == 45
+        assert e1.elevation == 35.264389682754654
+
+        # Dummy test
+        e2 = self.Magnetic(p1)
+        assert e2.xtype == 'magnetic'
+        assert e2.points.shape == (5, 3)
+
+    def test_to_from_dict(self):
+        p1 = np.array([[-10, -10, -10], [10, 10, 10]])
+        e1 = electrodes.Dipole(p1)
+        e2 = e1.copy()
+        assert e1 == e2
+
+    def test_basic_repr(self):
+        p1 = (100, 2, 6, 0, 90)
+        e1 = electrodes.Dipole(p1)
+        r1 = e1.__repr__()
+        assert 'Dipole' in r1
+        assert 'center={100.0; 2.0; 6.0} m' in r1
+        assert '0.0°' in r1
+        assert '90.0°' in r1
+        assert 'l=1.0 m' in r1
+
+        p2 = np.array([[-10, -10, -10], [10, 10, 10]])
+        e2 = electrodes.Dipole(p2)
+        r2 = e2.__repr__()
+        assert 'Dipole' in r2
+        assert 'e1={-10.0; -10.0; -10.0} m' in r2
+        assert 'e2={10.0; 10.0; 10.0} m' in r2
+
+    def test_warnings(self):
+
+        with pytest.raises(ValueError, match="The two electrodes are identic"):
+            electrodes.Dipole((0, 0, 0, 0, 0, 0))
+
+        with pytest.raises(ValueError, match="The two electrodes are identic"):
+            electrodes.Dipole([[0, 0, 0], [0, 0, 0]])
+
+        with pytest.raises(ValueError, match="Coordinates are wrong defined."):
+            electrodes.Dipole((0, 1, 2))
+
+        with pytest.raises(ValueError, match="Coordinates are wrong defined."):
+            electrodes.Dipole(np.zeros((2, 4)))
+
+        with pytest.raises(ValueError, match="Coordinates are wrong defined."):
+            electrodes.Dipole(np.zeros((5, 3)))
+
+
+def test_source():
+    p1 = [[0, 0, 0], [1, 0, 0]]
+    strength = np.pi
+    freq = 1.234
+
+    s1 = electrodes.Source(strength, coordinates=p1)
+    assert s1.strength == strength
+
+    grid = emg3d.TensorMesh([[1, 1], [1, 1], [1, 1]], (0, 0, 0))
+    sfield = emg3d.fields.get_source_field(grid, s1, freq)
+    assert s1.get_field(grid, freq) == sfield
+
+
+def test_tx_electric_dipole():
     s1a = electrodes.TxElectricDipole(
             (0, 0, 0, 45, 45), strength=np.pi, length=4)
+    assert s1a.xtype == 'electric'
     s1b = electrodes.TxElectricDipole.from_dict(s1a.to_dict())
     assert s1a == s1b
     s2a = electrodes.TxElectricDipole(
@@ -37,40 +210,58 @@ def test_inst_to_from_dict():
     assert_allclose(s1b.length, s2b.length)
     assert_allclose(s1b.length, s3b.length)
 
-    # TxMagneticDipole - 3 formats
-    s4a = electrodes.TxElectricDipole(
+
+def test_tx_magnetic_dipole():
+    s4a = electrodes.TxMagneticDipole(
             (0, 0, 0, 45, 45), strength=np.pi, length=4)
-    s4b = electrodes.TxElectricDipole.from_dict(s4a.to_dict())
+    assert s4a.xtype == 'magnetic'
+    s4b = electrodes.TxMagneticDipole.from_dict(s4a.to_dict())
     assert s4a == s4b
-    s5a = electrodes.TxElectricDipole(
+    s5a = electrodes.TxMagneticDipole(
             (-1, 1, -1, 1, -np.sqrt(2), np.sqrt(2)), strength=np.pi)
-    s5b = electrodes.TxElectricDipole.from_dict(s5a.to_dict())
+    s5b = electrodes.TxMagneticDipole.from_dict(s5a.to_dict())
     assert s5a == s5b
-    s6a = electrodes.TxElectricDipole(
+    s6a = electrodes.TxMagneticDipole(
             [[-1, -1, -np.sqrt(2)], [1, 1, np.sqrt(2)]], strength=np.pi)
-    s6b = electrodes.TxElectricDipole.from_dict(s6a.to_dict())
+    s6b = electrodes.TxMagneticDipole.from_dict(s6a.to_dict())
     assert s6a == s6b
     assert_allclose(s4b.points, s5b.points)
     assert_allclose(s4b.points, s6b.points)
     assert_allclose(s4b.strength, s5b.strength)
     assert_allclose(s4b.strength, s6b.strength)
-    assert_allclose(s4b.length, s5b.length)
-    assert_allclose(s4b.length, s6b.length)
+    assert_allclose(s4b.length*2, s5b.length)
+    assert_allclose(s4b.length*2, s6b.length)
 
-    # TxElectricWire
+
+def test_tx_electric_wire():
     n = np.sqrt(2)/2
     coo = np.array([[0, -n, 0], [n, 0, 0], [0, n, 0],
                     [-n, 0, n], [0, 0, -n]])
     s7a = electrodes.TxElectricWire(coo, np.pi)
+    assert s7a.xtype == 'electric'
     s7b = electrodes.TxElectricWire.from_dict(s7a.to_dict())
     assert s7a == s7b
     assert_allclose(s7a.center, 0)
 
 
+def test_rx_electric_point():
+    r1a = electrodes.RxElectricPoint((1200, -56, 23.214, 368, 15))
+    assert r1a.xtype == 'electric'
+    r1b = electrodes.RxElectricPoint.from_dict(r1a.to_dict())
+    assert r1a == r1b
+
+
+def test_rx_magnetic_point():
+    r2a = electrodes.RxMagneticPoint((-1200, 56, -23.214, 0, 90))
+    assert r2a.xtype == 'magnetic'
+    r2b = electrodes.RxMagneticPoint.from_dict(r2a.to_dict())
+    assert r2a == r2b
+
+
 def test_point_to_dipole():
     source = (10, 100, -1000, 0, 0)
     length = 111.0
-    out = electrodes._point_to_dipole(source, length)
+    out = electrodes.point_to_dipole(source, length)
     assert out.shape == (2, 3)
     assert out[0, 0] == source[0]-length/2
     assert out[1, 0] == source[0]+length/2
@@ -81,7 +272,7 @@ def test_point_to_dipole():
 
     source = (10, 100, -1000, 30, 60)
     length = 2.0
-    out = electrodes._point_to_dipole(source, length)
+    out = electrodes.point_to_dipole(source, length)
     assert_allclose(
         out,
         [[9.5669873, 99.75, -1000.8660254], [10.4330127, 100.25, -999.1339746]]
@@ -142,13 +333,13 @@ class TestDipoleToPoint:
             coo = np.array([[0, 0, 0], points])
 
             # 1.a Check angles degree
-            azm, elv, rad = electrodes._dipole_to_point(coo)
+            azm, elv, rad = electrodes.dipole_to_point(coo)
             assert_allclose(azm, values[0])
             assert_allclose(elv, values[1])
             assert_allclose(rad, np.linalg.norm(points))
 
             # 1.b Check angles radians
-            azm, elv, _ = electrodes._dipole_to_point(coo, deg=False)
+            azm, elv, _ = electrodes.dipole_to_point(coo, deg=False)
             assert_allclose(azm, values[2])
             assert_allclose(elv, values[3])
 
@@ -157,29 +348,29 @@ class TestDipoleToPoint:
             length = 2*np.linalg.norm(points)
 
             # 2.a Check points degree
-            coo = electrodes._point_to_dipole(
+            coo = electrodes.point_to_dipole(
                     (0, 0, 0, values[0], values[1]), length)
             assert_allclose(coo[1, :], points)
 
             # 2.b Check points radians
-            coo = electrodes._point_to_dipole(
+            coo = electrodes.point_to_dipole(
                     (0, 0, 0, values[2], values[3]), length, deg=False)
             assert_allclose(coo[1, :], points, atol=1e-15)
 
     def test_arbitrary(self):
 
         i_azm1, i_elv1, i_rad1 = -25, 88, 1e6
-        coords1 = electrodes._point_to_dipole(
+        coords1 = electrodes.point_to_dipole(
                 (1e6, 1e-6, 10, i_azm1, i_elv1), i_rad1)
-        o_azm1, o_elv1, o_rad1 = electrodes._dipole_to_point(coords1)
+        o_azm1, o_elv1, o_rad1 = electrodes.dipole_to_point(coords1)
         assert_allclose(i_azm1, o_azm1)
         assert_allclose(i_elv1, o_elv1)
         assert_allclose(i_rad1, o_rad1)
 
         i_azm2, i_elv2, i_rad2 = -33.3, 90, 2  # <= azm != 0.0
-        coords2 = electrodes._point_to_dipole(
+        coords2 = electrodes.point_to_dipole(
                 (1e6, -50, 3.33, i_azm2, i_elv2), i_rad2)
-        o_azm2, o_elv2, o_rad2 = electrodes._dipole_to_point(coords2)
+        o_azm2, o_elv2, o_rad2 = electrodes.dipole_to_point(coords2)
         assert_allclose(0.0, o_azm2)  # <= azm == 0.0
         assert_allclose(i_elv2, o_elv2)
         assert_allclose(i_rad2, o_rad2)
@@ -188,7 +379,7 @@ class TestDipoleToPoint:
 def test_point_to_square_loop():
     source = (10, 100, -1000, 45, 90)
     length = 4
-    out = electrodes._point_to_square_loop(source, length)
+    out = electrodes.point_to_square_loop(source, length)
     assert out.shape == (5, 3)
     assert_allclose(out[:, 0], [9, 9, 11, 11, 9])
     assert_allclose(out[:, 1], [101, 99, 99, 101, 101])
@@ -196,7 +387,7 @@ def test_point_to_square_loop():
 
     source = (0, 0, 0, 0, 0)
     length = 8
-    out = electrodes._point_to_square_loop(source, length)
+    out = electrodes.point_to_square_loop(source, length)
     assert out.shape == (5, 3)
     assert_allclose(out[:, 0], 0)
     assert_allclose(out[:, 1], [2, 0, -2, 0, 2])
@@ -204,7 +395,7 @@ def test_point_to_square_loop():
 
     source = (10, 100, -1000, 30, 60)
     length = 4
-    out = electrodes._point_to_square_loop(source, length)
+    out = electrodes.point_to_square_loop(source, length)
     assert_allclose(out[:, 0],
                     [9.292893, 8.93934, 10.707107, 11.06066, 9.292893])
     assert_allclose(out[:, 2],
@@ -213,28 +404,28 @@ def test_point_to_square_loop():
 
 
 def test_rotation():
-    assert_allclose(electrodes._rotation(0, 0), [1, 0, 0])
-    assert_allclose(electrodes._rotation(180, 0), [-1, 0, 0])
-    assert_allclose(electrodes._rotation(90, 0), [0, 1, 0])
-    assert_allclose(electrodes._rotation(-90, 0), [0, -1, 0])
-    assert_allclose(electrodes._rotation(0, 90), [0, 0, 1])
-    assert_allclose(electrodes._rotation(0, -90), [0, 0, -1])
+    assert_allclose(electrodes.rotation(0, 0), [1, 0, 0])
+    assert_allclose(electrodes.rotation(180, 0), [-1, 0, 0])
+    assert_allclose(electrodes.rotation(90, 0), [0, 1, 0])
+    assert_allclose(electrodes.rotation(-90, 0), [0, -1, 0])
+    assert_allclose(electrodes.rotation(0, 90), [0, 0, 1])
+    assert_allclose(electrodes.rotation(0, -90), [0, 0, -1])
 
     dazm, delv = 30, 60
     razm, relv = np.deg2rad(dazm), np.deg2rad(delv)
     assert_allclose(
-        electrodes._rotation(dazm, delv),
+        electrodes.rotation(dazm, delv),
         [np.cos(razm)*np.cos(relv), np.sin(razm)*np.cos(relv), np.sin(relv)])
 
     dazm, delv = -45, 180
     razm, relv = np.deg2rad(dazm), np.deg2rad(delv)
     assert_allclose(
-        electrodes._rotation(dazm, delv),
+        electrodes.rotation(dazm, delv),
         [np.cos(razm)*np.cos(relv), np.sin(razm)*np.cos(relv), np.sin(relv)],
         atol=1e-14)
 
     # Radians
     azm, elv = np.pi/3, np.pi/4
-    rot1 = electrodes._rotation(azm, elv, deg=False)
-    rot2 = electrodes._rotation(np.rad2deg(azm), np.rad2deg(elv))
+    rot1 = electrodes.rotation(azm, elv, deg=False)
+    rot2 = electrodes.rotation(np.rad2deg(azm), np.rad2deg(elv))
     assert_allclose(rot1, rot2)