+++ CHANGELOG.rst
@@ -37,7 +37,8 @@ Detailed changes
   positive upwards (anticlockwise from the horizontal plane).
 - Bugfix of the loop area for a magnetic dipole (the area was previously wrong 
   except for dipole length of 1).
-- Source strength cannot be zero.
+- Zero source strength does no longer mean "normalized", it means zero
+  strength (hence no source).
 - New class ``TxElectricWire`` for an arbitrary electric wire.